Deleted firestore and refactored folders
