+++ instantwords/lib/authentication/login.dart
@@ -1,0 +1,164 @@
+part of '../main.dart';
+
+class LoginPage extends StatefulWidget {
+  final FireStorage _storage;
+  final SpeechToTextProvider _speechProvider;
+  final translator;
+
+  LoginPage(this._storage, this._speechProvider, this.translator);
+
+  @override
+  State<LoginPage> createState() => new _LoginPageState();
+}
+
+class _LoginPageState extends State<LoginPage> {
+  final TextEditingController _emailFilter = new TextEditingController();
+  final TextEditingController _passwordFilter = new TextEditingController();
+
+  String _email = "";
+  String _password = "";
+  bool isInProgress = false;
+
+  _LoginPageState() {
+    _emailFilter.addListener(_emailListen);
+    _passwordFilter.addListener(_passwordListen);
+  }
+
+  void _emailListen() {
+    if (_emailFilter.text.isEmpty) {
+      _email = "";
+    } else {
+      _email = _emailFilter.text;
+    }
+  }
+
+  void _passwordListen() {
+    if (_passwordFilter.text.isEmpty) {
+      _password = "";
+    } else {
+      _password = _passwordFilter.text;
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return SafeArea(
+      child: ModalProgressHUD(
+        inAsyncCall: isInProgress,
+        child: new Scaffold(
+          appBar: _buildBar(context),
+          body: new Container(
+            padding: EdgeInsets.all(16.0),
+            child: new Column(
+              children: <Widget>[
+                _buildTextFields(),
+                _buildButtons(),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildBar(BuildContext context) {
+    return new AppBar(
+      title: new Text("InstantWords Login"),
+      centerTitle: true,
+    );
+  }
+
+  Widget _buildTextFields() {
+    return new Container(
+      child: new Column(
+        children: <Widget>[
+          new Container(
+            child: new TextField(
+              controller: _emailFilter,
+              keyboardType: TextInputType.emailAddress,
+              decoration: new InputDecoration(labelText: 'Email'),
+            ),
+          ),
+          new Container(
+            child: new TextField(
+              controller: _passwordFilter,
+              keyboardType: TextInputType.visiblePassword,
+              decoration: new InputDecoration(labelText: 'Password'),
+              obscureText: true,
+            ),
+          )
+        ],
+      ),
+    );
+  }
+
+  Widget _buildButtons() {
+    return new Container(
+      child: new Column(
+        children: <Widget>[
+          new RaisedButton(
+            child: new Text('Login'),
+            onPressed: _loginPressed,
+          ),
+          new FlatButton(
+            child: new Text('Dont have an account? Tap here to register.'),
+            onPressed: () {
+              Navigator.pushReplacement(
+                  context,
+                  MaterialPageRoute(
+                      builder: (context) => RegisterPage(widget._storage,
+                          widget._speechProvider, widget.translator)));
+            },
+          ),
+          new FlatButton(
+            child: new Text('Forgot Password?'),
+            onPressed: _passwordReset,
+          )
+        ],
+      ),
+    );
+  }
+
+  // These functions can self contain any user auth logic required, they all have access to _email and _password
+
+  void _loginPressed() async {
+    setState(() {
+      isInProgress = true;
+    });
+    print('The user wants to login with $_email and $_password');
+    final status =
+        await FireAuth().loginAccount(email: _email, password: _password);
+    setState(() {
+      isInProgress = false;
+    });
+    if (status == AuthResultStatus.successful) {
+      Navigator.pushReplacement(
+        context,
+        MaterialPageRoute(
+            builder: (context) => Dashboard(
+                widget._storage, widget._speechProvider, widget.translator)),
+      );
+    } else {
+      final errorMsg = AuthExceptionHandler.generateExceptionMessage(status);
+      _showAlertDialog(errorMsg);
+    }
+  }
+
+  _showAlertDialog(errorMsg) {
+    return showDialog(
+        context: context,
+        builder: (context) {
+          return AlertDialog(
+            title: Text(
+              'Login Failed',
+              style: TextStyle(color: Colors.black),
+            ),
+            content: Text(errorMsg),
+          );
+        });
+  }
+
+  void _passwordReset() {
+    print("The user wants a password reset request sent to $_email");
+  }
+}
